Test with i128

+++ increment_overflow_checks/src/lib.rs
@@ -8,13 +8,13 @@ pub struct IncrementContract;
 #[contractimpl]
 impl IncrementContract {
     /// Increment increments an internal counter, and returns the value.
-    pub fn increment(env: Env, n: u32) -> u32 {
+    pub fn increment(env: Env, n: i128) -> i128 {
         // Get the current count.
-        let mut count: u32 = env
+        let mut count: i128 = env
             .storage()
             .get(&COUNTER)
             .unwrap_or(Ok(0)) // If no value set, assume 0.
-            .unwrap(); // Panic if the value of COUNTER is not u32.
+            .unwrap(); // Panic if the value of COUNTER is not i128.
         log!(&env, "count: {}", count);
 
         // Increment the count.
@@ -28,13 +28,13 @@ impl IncrementContract {
     }
 
     /// Decrement decrements an internal counter, and returns the value.
-    pub fn decrement(env: Env, n: u32) -> u32 {
+    pub fn decrement(env: Env, n: i128) -> i128 {
         // Get the current count.
-        let mut count: u32 = env
+        let mut count: i128 = env
             .storage()
             .get(&COUNTER)
             .unwrap_or(Ok(0)) // If no value set, assume 0.
-            .unwrap(); // Panic if the value of COUNTER is not u32.
+            .unwrap(); // Panic if the value of COUNTER is not i128.
         log!(&env, "count: {}", count);
 
         // Decrement the count.
